memory: correct type of MemorySizeInKB
This is needed for this code to work on both 32 and 64 bits system.
According to the interface,
https://github.com/openbmc/phosphor-dbus-interfaces/blob/388b58f9a878f45b8ec243152cac5eb44ec90ced/yaml/xyz/openbmc_project/Inventory/Item/Dimm.interface.yaml#L9,
MemorySizeInKB is of type size_t.

Tested:
On real hardware, the memory resource is working as expected.

+++ redfish-core/lib/memory.hpp
@@ -465,8 +465,8 @@ inline void getDimmDataByService(std::shared_ptr<bmcweb::AsyncResp> aResp,
                 }
                 else if (property.first == "MemorySizeInKB")
                 {
-                    const uint32_t* memorySize =
-                        std::get_if<uint32_t>(&property.second);
+                    const size_t* memorySize =
+                        std::get_if<size_t>(&property.second);
                     if (memorySize == nullptr)
                     {
                         // Important property not in desired type
